Add global CSS reset and font styling
Introduced a CSS reset to standardize and normalize styling across all browsers. This helps in avoiding inconsistencies and minor differences in layout rendering. Besides resetting, added a default font 'Satoshi' to ensure consistent typography across the application. This was basically done to ensure a cleaner and more predictable starting point when styling elements.