Remove placeholder content from App Component
This commit removes large block of unnecessary placeholder content from the App Component. This content was initially added by the Angular CLI as a template for a generated application. Removal of this content will declutter the component, leaving only essential and useful code for a cleaner, easier to read file.