Add AppPaths enum and update routing
Implemented AppPaths enum to centralize app routings in one place. Enum AppPaths has been created in 'core/enums' directory and 'payments' route was added in it. Further, the app-routing module was updated to reflect these changes where 'payments' was mapped to the PaymentsComponent. This improves code readability and ease of future path management.