Add LOGIN to AppPaths and use in routing
In the enum AppPaths, the LOGIN route was added to keep all routes in one place. The path 'auth/login' was replaced with AppPaths.LOGIN in the routing module to improve consistency and maintainability. Now, if the login route needs to be changed, it should be done in one place: AppPaths enum.