Update Paginator styling and functionality
The changes in this commit involve the redesign of the Paginator component UI and an addition to its functionality. The component is now wrapped in a div flex box and it includes a record count along with left and right navigation icons, improving its aesthetics. The update on the app-paginator, reducing its length from 100 to 60 while increasing its pageSize from 9 to 20, aids with the improved performance and user interaction. An output event emitter has also been added to the paginator.component.ts, emitting the current page upon page change, enhancing the component’s interactability.