Add PayService and corresponding tests
Created PayService in order to retrieve and manage payments from a provided API. Implemented the getPayments function which fetches payments according to the provided skip and limit parameters. This service allows the app to retrieve payment data effectively. In addition, a corresponding service test file was created to ensure the reliability of the PayService. Lastly, established the PayResponse and Pay interfaces to handle the responses from the API correctly.