"Update color scheme in status-nav component"
The color for the text in the status-nav component has been updated. Previously, the color was hardcoded as #171717. Now, it has been changed to #707070 for standard text and a CSS variable (--primary-text) has been introduced for hovered and active states. This change allows for easier theme adjustability in future design updates.