Update text color to use CSS variable
The color of the texts were changed from a fixed hexadecimal color (#242424) to a dynamic CSS variable (--primary-text). This allows us to update the color associated with --primary-text at a single place in the future without changing individual SCSS files. This change provides better maintainability and consistency throughout the application. Deleting the --primary color from text-info-small and h2, since it was not serving any purpose. Added '.input-container-search' class structure for future use.