Update sidenav UI and enhance responsiveness
The sidenav component was updated for improving UI and enhancing responsiveness. The image source was modified using ngSrc rather than src for a more efficient loading. Moreover, icons, menu items, and profiles were restructured with additional classes for better layout management. The height of the host was set to 100% to ensure that the sidenav covers the full height regardless of the content.