Update user photo upload method
The method of photo uploading in user profile has been updated. Instead of selecting a file directly from the file system, the user is now shown a dialog for uploading their photo. This was changed to enhance the user experience by providing a more intuitive, interactive way of updating the profile picture. Components were also updated to reflect this change in the UI.