Update upload component for proper image handling
Reformatted import statements for better readability and ease of navigation. Changed the DialogRef type from File to String to pass image URLs instead of whole files, enhancing overall performance. Replaced the 'src' attribute in img tag with 'ngSrc' for better Angular compatibility and reliability. Also updated the dialog closing method to pass the URL instead of the file. These changes are to ensure more efficient and reliable handling of image uploads in the users' feature.