Add directive to block minus sign in OTP input fields
A new Angular directive (app-block-minus-sign) has been added to block minus signs on on-screen input fields. This was necessary to prevent negative numeral inputs, specifically in One-time-Password (OTP) fields, thus improving the user input interface and the security aspect of the application. As part of the changes, the directive is used in the OTP input components which were also added to the auth components.