Update login component logic and animations
Updated the logic and animation for the login component. Replaced the old static routerLink on the login button with a reactive onSubmit function that incorporates form validation to ensure the email and password are valid before making a login request to the authentication service. Additionally, animation trigger was renamed for clarity. These changes improve user experience and code readability.